Extracted IManualEventListener to its own file
