+++ src/Impostor.Api/Events/IEventListener.cs
@@ -1,16 +1,6 @@
-﻿using System.Threading.Tasks;
-
-namespace Impostor.Api.Events
+﻿namespace Impostor.Api.Events
 {
     public interface IEventListener
     {
-    }
-
-    public interface IManualEventListener : IEventListener
-    {
-        public bool CanExecute<T>();
-
-        public ValueTask Execute(IEvent @event);
-        EventPriority Priority { get; set; }
     }
 }
