+++ src/Impostor.Api/Events/IManualEventListener.cs
@@ -1,0 +1,13 @@
+﻿using System.Threading.Tasks;
+
+namespace Impostor.Api.Events
+{
+    public interface IManualEventListener : IEventListener
+    {
+        public bool CanExecute<T>();
+
+        public ValueTask Execute(IEvent @event);
+
+        EventPriority Priority { get; set; }
+    }
+}
